pyproject.toml
[project]
name = "nasa-client-pro"
version = "0.3.0"
description = "Simple NASA API Client"
readme = "README.md"
authors = [
    { name = "Theophilus Siameh", email = "tsiameh@hardtanh.com" }
]
requires-python = ">=3.14"
dependencies = [
    "black>=26.3.1",
    "build>=1.4.2",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "tenacity>=9.1.4",
    "twine>=6.2.0",
]

[project.scripts]
nasa-client = "nasa_client_pro:main"

[build-system]
requires = ["uv_build>=0.11.1,<0.12.0"]
build-backend = "uv_build"
